Fetch comments by post slug, clearify post / comment components name, add tags display
